+++ RELEASE_NOTES.md
@@ -12,6 +12,7 @@
 #### Items
 - Changed Quiver use action to equip it, rather than empty its contents
 - Changed Dagger to be able to parry swords in blade clashes
+- Changed Trident animation to use the Stab animation
 
 #### Blocks
 - Changed Iron Grates to have the `dragon_immune` block tag